backend/pyproject.toml
[tool.poetry]
name = "innoliber-backend"
version = "0.1.0"
description = "InnoLiber Backend API Service"
authors = ["InnoLiber Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.118.2"
uvicorn = {extras = ["standard"], version = "^0.31.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.35"}
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
pydantic-settings = "^2.6.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.0.0"  # 密码哈希：直接使用bcrypt而非passlib
python-multipart = "^0.0.12"
alembic = "^1.13.3"
redis = "^5.2.0"
celery = "^5.4.0"
torch = "^2.5.1"
openai = "^1.55.0"
httpx = "^0.27.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
mypy = "^1.13.0"
ipython = "^8.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"